refactor: move vitals related schemas into file
