+++ src/server/routers/eyesight_router.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 import { eq } from "drizzle-orm";
 import { router, protectedProcedure } from "@/server/trpc";
 import { db } from "@/db/drizzle";
-import { eyesight } from "@/db/schema/schema";
+import { eyesight } from "@/db/schema/vitals";
 
 /**
  * Common select fields for eyesight queries to ensure consistency
